# Conflicts:
#	AppCUI
#	GViewCore/src/View/BufferViewer/Instance.cpp
#	Types/MachOFB/include/Machofb.hpp
#	Types/MachOFB/src/CMakeLists.txt
#	Types/MachOFB/src/MachOFBFile.cpp
#	Types/MachOFB/src/Machofb.cpp
#	Types/UniversalMachO/src/Panels/Information.cpp
#	Types/UniversalMachO/src/Panels/Objects.cpp

+++ GViewCore/src/View/BufferViewer/Instance.cpp
@@ -100,7 +100,7 @@ void Instance::OpenCurrentSelection()
     {
         LocalString<128> temp;
         temp.Format("Buffer_%llx_%llx", start, end);
-        GView::App::OpenItem((ExtractItem) (res), this, (end - start) + 1, temp);
+        GView::App::OpenItem((ExtractItem) (size_t) (res), this, (end - start) + 1, temp);
     }
 }
 void Instance::UpdateCurrentSelection()
